+++ trunk/Source/C++/Core/Ap4Expandable.cpp
@@ -117,8 +117,8 @@ AP4_Expandable::Inspect(AP4_AtomInspector& inspector)
     AP4_FormatString(name, sizeof(name), "[#:%02x]", m_ClassId);
     char info[64];
     AP4_FormatString(info, sizeof(info), "size=%d+%d",
-                     GetHeaderSize(),
-                     m_PayloadSize);
+                     (int)GetHeaderSize(),
+                     (int)m_PayloadSize);
     inspector.StartElement(name, info);
     inspector.EndElement();
