+++ trunk/Source/C++/Core/Ap4ObjectDescriptor.cpp
@@ -156,7 +156,7 @@ AP4_ObjectDescriptor::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
     AP4_FormatString(info, sizeof(info), "size=%d+%d", 
-                     GetHeaderSize(),m_PayloadSize);
+                     (int)GetHeaderSize(),(int)m_PayloadSize);
     inspector.StartElement("[ObjectDescriptor]", info);
     inspector.AddField("id", m_ObjectDescriptorId);
     if (m_UrlFlag) {
@@ -306,7 +306,7 @@ AP4_InitialObjectDescriptor::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
     AP4_FormatString(info, sizeof(info), "size=%d+%d", 
-                     GetHeaderSize(),m_PayloadSize);
+                     (int)GetHeaderSize(),(int)m_PayloadSize);
     inspector.StartElement("[InitialObjectDescriptor]", info);
     inspector.AddField("id", m_ObjectDescriptorId);
     if (m_UrlFlag) {
@@ -388,7 +388,7 @@ AP4_DescriptorUpdateCommand::Inspect(AP4_AtomInspector& inspector)
 {
     char info[64];
     AP4_FormatString(info, sizeof(info), "size=%d+%d", 
-                     GetHeaderSize(),m_PayloadSize);
+                     (int)GetHeaderSize(),(int)m_PayloadSize);
     switch (GetTag()) {
         case AP4_COMMAND_TAG_OBJECT_DESCRIPTOR_UPDATE:
             inspector.StartElement("[ObjectDescriptorUpdate]", info);
